migrate to `tower-sessions` (#104)
We have entirely reimagined this crate's API, having now rewritten it to use `tower-sessions`.

The upshot of this is issues with deadlocks are a thing of the past. However, applications that rely on prior versions of the crate will have to evaluate the new API and decide if it's appropriate to migrate or not.