ensure anonymous sessions are verified
This patch differentiates between sessions that have already been
authenticated and those that have not when verifying a session. This
distinction is important because when a logged in user's session cannot
be verified we must flush the session. However, before a user is logged
in we should avoid unnecessarily removing session data.