fix integration tests
For one reason or another, the ability to rely on the set-cookie header
to introspect cookies via reqwest seems to no longer work with Github
Actions; this otherwise works in a local environment so the reason is
not immediately clear.

That said this can be worked around by using a cookie jar directly and
observing the changes the client is making therein. This fix is applied
here.

At the same time a few minor clean ups are provided.