refactor(session): ensure thread-safe state
Originally the auth session was not designed to be shared between
middleware. However, in practice it's convenient and indeed idiomatic
to share auth between middleware before a request has been fully
resolved. This allows e.g. specialized handling of auth semantics,
such that applications may leverage the session indirectly.

To do so, the entire state of the auth session must be reworked to
support interior mutability. This allows the session state to be passed
between and manipulated by intermediary middleware. The downside to this
approach is that locking is required given the fundamental nature of
middleware. But the benefit is that now state is visible where it
otherwise was not.

This also introduces several breaking changes:

1. Users are now accessible via the `user` method, which must be
   awaited.
2. The backend is no longer an attribute and instead must be accessed
   through the `backend` method.
3. Access to the session has been removed since this requires a clone;
   instead the `Session` should simply be used directly.

The examples and tests have been updated to reflect these changes.